+++ src/aks-preview/azext_aks_preview/addonconfiguration.py
@@ -4,12 +4,11 @@
 # --------------------------------------------------------------------------------------------
 
 import json
-import datetime
 from knack.log import get_logger
 from knack.util import CLIError
 from azure.cli.core.azclierror import ArgumentUsageError, ClientRequestError
 from azure.cli.core.commands import LongRunningOperation
-from azure.cli.core.commands.client_factory import get_subscription_id, get_mgmt_service_client
+from azure.cli.core.commands.client_factory import get_subscription_id
 from azure.cli.core.util import sdk_no_wait
 from azext_aks_preview._client_factory import CUSTOM_MGMT_AKS_PREVIEW
 from ._client_factory import cf_resources, cf_resource_groups
@@ -664,121 +663,6 @@ def ensure_container_insights_for_monitoring(cmd,
                     error = e
             else:
                 raise error
-
-    else:
-        # legacy auth with LA workspace solution
-        unix_time_in_millis = int(
-            (datetime.datetime.utcnow() - datetime.datetime.utcfromtimestamp(0)).total_seconds() * 1000.0)
-
-        solution_deployment_name = 'ContainerInsights-{}'.format(
-            unix_time_in_millis)
-
-        # pylint: disable=line-too-long
-        template = {
-            "$schema": "https://schema.management.azure.com/schemas/2015-01-01/deploymentTemplate.json#",
-            "contentVersion": "1.0.0.0",
-            "parameters": {
-                "workspaceResourceId": {
-                    "type": "string",
-                    "metadata": {
-                        "description": "Azure Monitor Log Analytics Resource ID"
-                    }
-                },
-                "workspaceRegion": {
-                    "type": "string",
-                    "metadata": {
-                        "description": "Azure Monitor Log Analytics workspace region"
-                    }
-                },
-                "solutionDeploymentName": {
-                    "type": "string",
-                    "metadata": {
-                        "description": "Name of the solution deployment"
-                    }
-                }
-            },
-            "resources": [
-                {
-                    "type": "Microsoft.Resources/deployments",
-                    "name": "[parameters('solutionDeploymentName')]",
-                    "apiVersion": "2017-05-10",
-                    "subscriptionId": "[split(parameters('workspaceResourceId'),'/')[2]]",
-                    "resourceGroup": "[split(parameters('workspaceResourceId'),'/')[4]]",
-                    "properties": {
-                        "mode": "Incremental",
-                        "template": {
-                            "$schema": "https://schema.management.azure.com/schemas/2015-01-01/deploymentTemplate.json#",
-                            "contentVersion": "1.0.0.0",
-                            "parameters": {},
-                            "variables": {},
-                            "resources": [
-                                {
-                                    "apiVersion": "2015-11-01-preview",
-                                    "type": "Microsoft.OperationsManagement/solutions",
-                                    "location": "[parameters('workspaceRegion')]",
-                                    "name": "[Concat('ContainerInsights', '(', split(parameters('workspaceResourceId'),'/')[8], ')')]",
-                                    "properties": {
-                                        "workspaceResourceId": "[parameters('workspaceResourceId')]"
-                                    },
-                                    "plan": {
-                                        "name": "[Concat('ContainerInsights', '(', split(parameters('workspaceResourceId'),'/')[8], ')')]",
-                                        "product": "[Concat('OMSGallery/', 'ContainerInsights')]",
-                                        "promotionCode": "",
-                                        "publisher": "Microsoft"
-                                    }
-                                }
-                            ]
-                        },
-                        "parameters": {}
-                    }
-                }
-            ]
-        }
-
-        params = {
-            "workspaceResourceId": {
-                "value": workspace_resource_id
-            },
-            "workspaceRegion": {
-                "value": location
-            },
-            "solutionDeploymentName": {
-                "value": solution_deployment_name
-            }
-        }
-
-        deployment_name = 'aks-monitoring-{}'.format(unix_time_in_millis)
-        # publish the Container Insights solution to the Log Analytics workspace
-        return _invoke_deployment(cmd, resource_group, deployment_name, template, params,
-                                  validate=False, no_wait=False, subscription_id=subscription_id)
-
-
-def _invoke_deployment(cmd, resource_group_name, deployment_name, template, parameters, validate, no_wait,
-                       subscription_id=None):
-    from azure.cli.core.profiles import ResourceType
-    DeploymentProperties = cmd.get_models(
-        'DeploymentProperties', resource_type=ResourceType.MGMT_RESOURCE_RESOURCES)
-    properties = DeploymentProperties(
-        template=template, parameters=parameters, mode='incremental')
-    smc = get_mgmt_service_client(cmd.cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES,
-                                  subscription_id=subscription_id).deployments
-    if validate:
-        logger.info('==== BEGIN TEMPLATE ====')
-        logger.info(json.dumps(template, indent=2))
-        logger.info('==== END TEMPLATE ====')
-
-    Deployment = cmd.get_models(
-        'Deployment', resource_type=ResourceType.MGMT_RESOURCE_RESOURCES)
-    deployment = Deployment(properties=properties)
-    if validate:
-        if cmd.supported_api_version(min_api='2019-10-01', resource_type=ResourceType.MGMT_RESOURCE_RESOURCES):
-            validation_poller = smc.begin_validate(
-                resource_group_name, deployment_name, deployment)
-            return LongRunningOperation(cmd.cli_ctx)(validation_poller)
-        else:
-            return smc.validate(resource_group_name, deployment_name, deployment)
-
-    return sdk_no_wait(no_wait, smc.begin_create_or_update, resource_group_name, deployment_name, deployment)
 
 
 def add_monitoring_role_assignment(result, cluster_resource_id, cmd):