+++ src/appservice-kube/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+0.1.4
+++++++
+* Ensure compatibility of 'az webapp create' and 'az functionapp create' with CLI version 2.34.0
+
+0.1.3
+++++++
+* Update functionapp runtimes and support v4 functionapps
+
 0.1.2
 ++++++
 * Allow passing custom locations by name if in the same resource group as the app/plan